# api/openapi.yaml
openapi: 3.1.0
info:
  title: API Quiz Bank
  version: 0.1.0
  summary: MVP API contract for governed quiz delivery.
servers:
  - url: http://127.0.0.1:8000
    description: Local MVP API endpoint.
  - url: https://api.valerchik.de
    description: Protected beta route; edge deployments may also require X-API-Key.
paths:
...
  /v1/quiz-items/next:
    post:
      operationId: getNextQuizItem
      summary: Select and log the next eligible quiz item for a governed consumer.
      tags:
        - quiz-delivery
      parameters:
        - name: X-Consumer-Id
          in: header
          required: true
          schema:
            type: string
            minLength: 1
        - name: X-QuizBank-API-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
        - name: X-QuizBank-Quota-Key
          in: header
          required: false
          description: Required for website_quiz_teaser; optional for shorts_factory_backend and regular server-to-server consumers.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NextQuizRequest'
      responses:
        '200':
          description: Eligible quiz item projection and delivery log metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextQuizResponse'
        '404':
          description: No eligible quiz item is available.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Missing or invalid consumer-bound API credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Consumer credential is inactive, unauthorized or not entitled.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Missing or invalid scoped quota key for consumers that require per-session quota.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Consumer quota or endpoint rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/quiz-items/{item_id}:
    get:
      operationId: getTrustedQuizItemById
      summary: Read an answer-enabled approved or published quiz item for a trusted consumer.
      tags:
        - quiz-delivery
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
        - name: X-Consumer-Id
          in: header
          required: true
          schema:
            type: string
            minLength: 1
        - name: X-QuizBank-API-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Answer-enabled quiz item projection for trusted manual quiz_id flow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedQuizItemLookupResponse'
        '401':
          description: Missing or invalid consumer-bound API credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Consumer is not trusted, active or entitled for this item.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Approved or published item is not visible for this trusted lookup.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
...
  schemas:
...
    NextQuizRequest:
      type: object
      additionalProperties: false
      required:
        - consumer_id
      properties:
        consumer_id:
          type: string
          minLength: 1
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        theme_ids:
          type: array
          items:
            type: string
            pattern: '^T(0[1-9]|1[0-8])$'
        objective_ids:
          type: array
          items:
            type: string
            pattern: '^O(0[1-9]|1[0-6])$'
        pattern_ids:
          type: array
          items:
            type: string
            pattern: '^P(0[1-9]|1[0-2])$'
    QuizItemPublicProjection:
      type: object
      additionalProperties: false
      required:
        - id
        - public_id
        - language
        - question
        - options
        - cefr_level
        - theme
        - objective
        - pattern
        - status
        - metadata
      properties:
        id:
          type: string
        public_id:
          type: string
        language:
          type: string
          const: de
        question:
          type: object
          additionalProperties: false
          required: [text, prompt, stem]
          properties:
            text:
              type: string
            prompt:
              type: string
            stem:
              type: string
        options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            type: object
            additionalProperties: false
            required: [id, position, text]
            properties:
              id:
                type: string
              position:
                type: integer
                minimum: 1
              text:
                type: string
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        theme:
          type: object
          additionalProperties: false
          required: [title, slug]
          properties:
            title:
              type: string
            slug:
              type: string
        objective:
          type: object
          additionalProperties: false
          required: [title]
          properties:
            title:
              type: string
        pattern:
          type: object
          additionalProperties: false
          required: [title]
          properties:
            title:
              type: string
        status:
          type: string
          enum: [approved, published]
        metadata:
          type: object
          additionalProperties: false
          required: [display]
          properties:
            display:
              type: object
              additionalProperties: false
              required:
                - cefr_level
                - theme_title
                - theme_slug
                - objective_title
                - pattern_title
              properties:
                cefr_level:
                  type: string
                theme_title:
                  type: string
                theme_slug:
                  type: string
                objective_title:
                  type: string
                pattern_title:
                  type: string
    QuizItemAnswerFeedback:
      type: object
      additionalProperties: false
      required: [correctAnswerId, explanation]
      properties:
        correctAnswerId:
          type: string
          description: Option id from quiz_item.options[].id.
        explanation:
          type: string
    QuizItemAnswerEnabledProjection:
      type: object
      additionalProperties: true
      required:
        - id
        - public_id
        - language
        - question
        - options
        - cefr_level
        - theme
        - objective
        - pattern
        - status
        - metadata
        - feedback
      properties:
        id:
          type: string
        public_id:
          type: string
        language:
          type: string
          const: de
        question:
          type: object
        options:
          type: array
          items:
            type: object
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        theme:
          type: object
        objective:
          type: object
        pattern:
          type: object
        status:
          type: string
          enum: [approved, published]
        metadata:
          type: object
        feedback:
          $ref: '#/components/schemas/QuizItemAnswerFeedback'
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, reason_code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        reason_code:
          type: string
...
    DeliveryProjection:
      type: object
      additionalProperties: false
      required:
        - delivery_id
        - consumer_id
        - quiz_item_id
        - item_status
        - status
        - selected_at
        - reason
      properties:
        delivery_id:
          type: string
        consumer_id:
          type: string
        quiz_item_id:
          type: string
        item_status:
          type: string
          enum: [approved, published]
        status:
          type: string
          enum: [created, reserved, delivered, sent, failed, skipped, expired, cancelled]
        selected_at:
          type: string
        reason:
          type: string
...
    TrustedQuizItemLookupResponse:
      type: object
      additionalProperties: false
      required: [consumer_id, quiz_item, interaction, access]
      properties:
        consumer_id:
          type: string
        quiz_item:
          $ref: '#/components/schemas/QuizItemAnswerEnabledProjection'
        interaction:
          type: object
          additionalProperties: false
          required: [mode, answer_key_included]
          properties:
            mode:
              type: string
              const: trusted_item_lookup
            answer_key_included:
              type: boolean
              const: true
        access:
          type: object
          additionalProperties: false
          required: [entitlement_checked, deliverable_status_checked]
          properties:
            entitlement_checked:
              type: boolean
            deliverable_status_checked:
              type: boolean
    NextQuizResponse:
      type: object
      additionalProperties: false
      required:
        - delivery_id
        - consumer_id
        - quiz_item
        - delivery
        - interaction
        - selection
      properties:
        delivery_id:
          type: string
        consumer_id:
          type: string
        quiz_item:
          oneOf:
            - $ref: '#/components/schemas/QuizItemPublicProjection'
            - $ref: '#/components/schemas/QuizItemAnswerEnabledProjection'
        delivery:
          $ref: '#/components/schemas/DeliveryProjection'
        interaction:
          type: object
          additionalProperties: false
          required: [mode, answer_key_included]
          properties:
            mode:
              type: string
            answer_key_included:
              type: boolean
        selection:
          type: object
          additionalProperties: false
          required: [entitlement_checked, quota_checked, repeat_policy_applied, decision]
          properties:
            entitlement_checked:
              type: boolean
            quota_checked:
              type: boolean
            repeat_policy_applied:
              type: boolean
            decision:
              $ref: '#/components/schemas/SelectionDecisionMetadata'
    SelectionDecisionMetadata:
      type: object
      additionalProperties: false
      required:
        - selection_request_id
        - candidate_count
        - eligible_count
        - selected_reason
        - selected_score
        - blocked_reason_counts
        - fallback_reason_code
      properties:
        selection_request_id:
          type: string
        candidate_count:
          type: integer
          minimum: 0
        eligible_count:
          type: integer
          minimum: 0
        selected_reason:
          type: string
        selected_score:
          type: object
          additionalProperties:
            type: number
        blocked_reason_counts:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
        fallback_reason_code:
          type:
            - string
            - 'null'